+++ source_gen/source_gen/lib/generators/json_serializable_generator.dart
@@ -2,14 +2,14 @@ library source_gen.json_serial.generator;
 
 import 'dart:async';
 import 'package:analyzer/src/generated/element.dart';
+import 'package:analyzer/src/generated/utilities_dart.dart';
 
 import 'package:source_gen/source_gen.dart';
 import 'package:source_gen/src/utils.dart';
 
 import 'json_serializable.dart';
 
-// TODO: assumes there is a empty, default ctor
-// TODO: assumes all fields are set-able
+// TODO: toJson option to omit null/empty values
 class JsonSerializableGenerator
     extends GeneratorForAnnotation<JsonSerializable> {
   const JsonSerializableGenerator();
@@ -27,45 +27,22 @@ class JsonSerializableGenerator
     var classElement = element as ClassElement;
     var className = classElement.name;
 
-    var finalFields = new Set<String>();
+    // Get all of the fields that need to be assigned
+    // TODO: support overriding the field set with an annotation option
+    var fields = classElement.fields.fold(<String, FieldElement>{},
+        (map, field) {
+      map[field.name] = field;
+      return map;
+    }) as Map<String, FieldElement>;
 
-    var fields = <String, String>{};
-    for (FieldElement field in classElement.fields) {
-      if (field.isFinal) {
-        finalFields.add(field.name);
-      }
-      fields[field.name] = field.type.name;
-    }
-
-    if (finalFields.isNotEmpty) {
-      throw new InvalidGenerationSourceError(
-          'Generator cannot target `$className`.',
-          todo: 'Make the following fields writable: ${finalFields.join(', ')}.');
-    }
+    // Get the constructor to use for the factory
 
     var prefix = '_\$${className}';
 
     var buffer = new StringBuffer();
 
     if (annotation.createFactory) {
-
-      //
-      // Generate the static factory method
-      //
-      buffer.writeln();
-      buffer
-          .writeln('$className ${prefix}FromJson(Map<String, Object> json) =>');
-      buffer.write('    new $className()');
-      if (fields.isEmpty) {
-        buffer.writeln(';');
-      } else {
-        fields.forEach((name, type) {
-          buffer.writeln();
-          buffer.write("      ..${name} = json['$name']");
-        });
-        buffer.writeln(';');
-      }
-      buffer.writeln();
+      _writeFactory(buffer, classElement, fields, prefix);
     }
 
     if (annotation.createToJson) {
@@ -75,16 +52,16 @@ class JsonSerializableGenerator
       buffer.writeln('abstract class ${prefix}SerializerMixin {');
 
       // write fields
-      fields.forEach((k, v) {
-        buffer.writeln('  $v get $k;');
+      fields.forEach((name, field) {
+        buffer.writeln('  ${field.type.name} get $name;');
       });
 
       // write toJson method
       buffer.writeln('  Map<String, Object> toJson() => {');
 
       var pairs = <String>[];
-      fields.forEach((k, v) {
-        pairs.add("'$k': $k");
+      fields.forEach((name, field) {
+        pairs.add("'$name': ${_fieldToJsonMapValue(name, field)}");
       });
       buffer.writeln(pairs.join(','));
 
@@ -98,4 +75,101 @@ class JsonSerializableGenerator
 
   @override
   String toString() => 'JsonGenerator';
+}
+
+void _writeFactory(StringBuffer buffer, ClassElement classElement,
+    Map<String, FieldElement> fields, String prefix) {
+  // creating a copy so it can be mutated
+  var fieldsToSet = new Map<String, FieldElement>.from(fields);
+  var className = classElement.displayName;
+  // Create the factory method
+
+  // Get the default constructor
+  // TODO: allow overriding the ctor used for the factory
+  var ctor = classElement.constructors.singleWhere((ce) => ce.name == '');
+
+  var ctorArguments = <String>[];
+  var ctorNamedArguments = <String>[];
+
+  for (var arg in ctor.parameters) {
+    var field = fields[arg.name];
+
+    if (field == null) {
+      if (arg.parameterKind == ParameterKind.REQUIRED) {
+        throw 'Cannot populate the required constructor argument: ${arg.displayName}.';
+      }
+      continue;
+    }
+
+    // TODO: validate that the types match!
+    if (arg.parameterKind == ParameterKind.NAMED) {
+      ctorNamedArguments.add(arg.name);
+    } else {
+      ctorArguments.add(arg.name);
+    }
+    fieldsToSet.remove(arg.name);
+  }
+
+  var finalFields = fieldsToSet.values.where((field) => field.isFinal).toSet();
+
+  if (finalFields.isNotEmpty) {
+    throw new InvalidGenerationSourceError(
+        'Generator cannot target `$className`.',
+        todo: 'Make the following fields writable or add them to the '
+          'constructor with matching names: '
+        '${finalFields.map((field) => field.name).join(', ')}.');
+  }
+
+  //
+  // Generate the static factory method
+  //
+  buffer.writeln();
+  buffer.writeln('$className ${prefix}FromJson(Map<String, Object> json) =>');
+  buffer.write('    new $className(');
+  buffer.writeAll(
+      ctorArguments.map((name) => _jsonMapAccessToField(name, fields[name])), ', ');
+  if (ctorArguments.isNotEmpty && ctorNamedArguments.isNotEmpty) {
+    buffer.write(', ');
+  }
+  buffer.writeAll(
+      ctorNamedArguments.map((name) => '$name: ' + _jsonMapAccessToField(name, fields[name])), ', ');
+
+  buffer.write(')');
+  if (fieldsToSet.isEmpty) {
+    buffer.writeln(';');
+  } else {
+    fieldsToSet.forEach((name, field) {
+      buffer.writeln();
+      buffer.write("      ..${name} = ");
+      buffer.write(_jsonMapAccessToField(name, field));
+    });
+    buffer.writeln(';');
+  }
+  buffer.writeln();
+}
+
+String _fieldToJsonMapValue(String name, FieldElement field) {
+  var result = name;
+
+  if (_isDartDateTime(field.type)) {
+    return "$name == null ? null : ${name}.toIso8601String()";
+  }
+
+  return result;
+}
+
+String _jsonMapAccessToField(String name, FieldElement field) {
+  var result = "json['$name']";
+
+  if (_isDartDateTime(field.type)) {
+    // TODO: this does not take into account that dart:core could be
+    // imported with another name
+    return "json.containsKey('$name') ? DateTime.parse($result) : null";
+  }
+
+  return result;
+}
+
+bool _isDartDateTime(DartType type) {
+  return type.element.library.isDartCore && type.name == 'DateTime';
 }